allow non-integer width/height in windows().
git-svn-id: https://svn.r-project.org/R/branches/R-1-2-patches@12991 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/gnuwin32/devga.c
@@ -1079,7 +1079,7 @@ static void devga_sbf(control c, int pos)
 
 
 static int 
-setupScreenDevice(DevDesc *dd, gadesc *xd, int w, int h, 
+setupScreenDevice(DevDesc *dd, gadesc *xd, double w, double h, 
 		  Rboolean recording, int resize) 
 {
     menu  m;
@@ -1087,8 +1087,8 @@ setupScreenDevice(DevDesc *dd, gadesc *xd, int w, int h,
     double dw, dw0, dh, d;
 
     xd->kind = SCREEN;
-    dw = dw0 = w / pixelWidth(NULL);
-    dh = h / pixelHeight(NULL);
+    dw = dw0 = (int) (w / pixelWidth(NULL));
+    dh = (int) (h / pixelHeight(NULL));
     if (resize != 3) {
 	if ((dw / devicewidth(NULL)) > 0.85) {
 	    d = dh / dw;
